requirements.txt
pyaml
htmlmin